refactor: main.js 정리 및 라우터/페이지/컴포넌트 구조 모듈화
