+++ src/components/footer.js
@@ -1,0 +1,6 @@
+// path: ~/Develop/front_5th_chapter1-1/src/components/footer.js
+export const Footer = () => `
+  <footer class="bg-gray-200 p-4 text-center">
+    <p>&copy; 2024 항해플러스. All rights reserved.</p>
+  </footer>
+`;
