+++ src/components/header.js
@@ -1,0 +1,6 @@
+// path: ~/Develop/front_5th_chapter1-1/src/components/header.js
+export const Header = () => `
+  <header class="bg-blue-600 text-white p-4 sticky top-0">
+    <h1 class="text-2xl font-bold">항해플러스</h1>
+  </header>
+`;
