+++ src/components/index.js
@@ -1,0 +1,3 @@
+export * from "./header.js";
+export * from "./navigation.js";
+export * from "./footer.js";
