+++ src/pages/index.js
@@ -1,0 +1,4 @@
+export * from "./mainPage.js";
+export * from "./profilePage.js";
+export * from "./loginPage.js";
+export * from "./errorPage.js";
