+++ src/utils/index.js
@@ -1,0 +1,3 @@
+// path: ~/Develop/front_5th_chapter1-1/src/utils/index.js
+
+export * from "./timestamp.js";
